move filtering to separate file
